KANA_VULN_SYSTEM/requirements.txt
aiohttp==3.11.10
aiosqlite==0.20.0
asyncpg==0.29.0
cryptography==44.0.0
loguru==0.7.2
python-dotenv==1.0.1
PyYAML==6.0.2
